+++ R/artefact_viewer.R
@@ -148,3 +148,5 @@ epochPlotly <- function(id,
     )
   )
 }
+
+