+++ R/dplyr-extensions.R
@@ -56,6 +56,40 @@ filter.eeg_data <- function(.data, ...) {
   .data
 }
 
+#' @importFrom dplyr filter
+#' @export
+filter.eeg_evoked <- function(.data,
+                              ...) {
+
+  orig_cols <- channel_names(.data)
+  args <- rlang::exprs(...)
+  .data$signals <- as.data.frame(.data)
+  .data$signals <- dplyr::filter(.data$signals,
+                                 ...)
+  .data$signals <- .data$signals[, orig_cols]
+  .data$timings <- dplyr::filter(.data$timings,
+                                 ...)
+  .data$events <- dplyr::filter(.data$events,
+                                ...)
+
+  #conditionally filter the epochs structure if any of the arguments refer to
+  #its contents. also need to fix this for timings - if we filter based on the
+  #epochs structure, it may miss out the timings structure
+  if (is.null(.data$epochs)) {
+    warning("Epochs structure missing; Update your eeg_epochs object using update_eeg_epochs.")
+    return(.data)
+  }
+
+  epo_args <- grepl(paste(names(.data$epochs), collapse = "|"),
+                    unlist(args))
+  if (any(epo_args)) {
+    .data$epochs <- dplyr::filter(.data$epochs,
+                                  !!!args[epo_args])
+  }
+  .data
+}
+
+
 #' @importFrom dplyr select
 #' @export
 dplyr::select
@@ -89,7 +123,15 @@ select.eeg_data <- function(.data, ...) {
 select.eeg_ICA <- function(.data, ...) {
   .data$signals <- dplyr::select(.data$signals, ...)
   .data$mixing_matrix <- dplyr::select(.data$mixing_matrix, ..., electrode)
-  .data$unmixing_matrix <- dplyr::select(.data$unmixing_matrix, ..., electrode)
+  unmix <- data.table::transpose(.data$unmixing_matrix[, 1:(ncol(.data$unmixing_matrix)-1)])
+  names(unmix) <- .data$unmixing_matrix$Component
+  unmix <- dplyr::select(unmix, ...)
+  remaining_comps <- names(unmix)
+  unmix <- data.table::transpose(unmix)
+  unmix$Component <- remaining_comps
+  names(unmix) <- names(.data$unmixing_matrix)
+  .data$unmixing_matrix <- unmix
+
   if (!is.null(.data$chan_info)) {
     .data$chan_info <- .data$chan_info[.data$chan_info$electrode %in% .data$mixing_matrix$electrode, ]
   }