+++ R/run_ICA.R
@@ -19,74 +19,135 @@ run_ICA <- function(data, ...) {
 #' @param maxit Maximum number of iterations of the Infomax and Fastica ICA
 #'   algorithms.
 #' @param tol Convergence tolerance for fastica and infomax. Defaults to 1e-06.
+#' @param pca Reduce the number of dimensions using PCA before running ICA.
+#'   Numeric,  >1 and < number of channels
 #' @describeIn run_ICA Run ICA on an \code{eeg_epochs} object
 #' @export
 
 run_ICA.eeg_epochs <- function(data,
                                method = "sobi",
-                               maxit = 500,
+                               maxit = 1000,
                                tol = 1e-6,
+                               pca = NULL,
                                ...) {
 
-  rank_check <- Matrix::rankMatrix(as.matrix(data$signals))
-  #if (rank_check < n_channels) {
-   # warning(paste("Data is rank deficient. Detected rank", rank_check))
-    #pca <- prcomp(data$signals)
-    #data$signals <- as.data.frame(pca$x[, 1:rank_check] %*% t(pca$rotation[, 1:rank_check]))
-  #}
+  if (!is.null(pca)) {
+    message("Reducing data to ", pca, " dimensions using PCA.")
+    orig_chans <- channel_names(data)
+    pca_decomp <- eigen(cov(data$signals))$vectors #prcomp(data$signals, scale = FALSE)
+    data$signals <- as.data.frame(as.matrix(data$signals) %*% pca_decomp[, 1:pca])
+    pca_flag <- TRUE
+  } else {
+    pca <- ncol(data$signals)
+    pca_flag <- FALSE
+  }
 
-  if (method == "sobi") {
-    if (!requireNamespace("JADE", quietly = TRUE)) {
-      stop("Package \"JADE\" needed to use SOBI. Please install it.",
+  rank_check <- Matrix::rankMatrix(as.matrix(data$signals))
+
+  if (rank_check < ncol(data$signals)) {
+    warning(paste("Data is rank deficient. Detected rank", rank_check))
+  }
+
+
+  if (!(method %in% c("sobi", "fastica", "infomax", "fica"))) {
+    stop("Unknown method; available methods are sobi, fastica, infomax, and fica.")
+  }
+
+  if (method == "fica") {
+
+    if (!requireNamespace("fICA", quietly = TRUE)) {
+      stop("Package \"fICA\" needed to use the fICA implementation of fastica. Please install it.",
            call. = FALSE)
     }
-    ica_obj <- sobi_ICA(data,
-                        maxiter = maxit)
+    message("Running fastica (fICA).")
+
+    ICA_out <- fICA::fICA(as.matrix(data$signals),
+                          eps = tol,
+                          maxiter = maxit,
+                          method = "sym2")
+
+    ICA_out$S <- tibble::as_tibble(ICA_out$S[, 1:pca])
+
+    ICA_out$W <- ICA_out$W[, 1:pca]
+    mixing_matrix <- MASS::ginv(ICA_out$W, tol = 0)
+    mixing_matrix <- pca_decomp[, 1:pca] %*% mixing_matrix
+
+    unmixing_matrix <- as.data.frame(MASS::ginv(mixing_matrix, tol = 0))
+
+    mixing_matrix <- as.data.frame(mixing_matrix)
+    names(ICA_out$S) <- paste0("Comp", 1:pca)
+
+    names(mixing_matrix) <- paste0("Comp", 1:pca)
+    mixing_matrix$electrode <- orig_chans #names(data$signals)
+    names(unmixing_matrix) <- orig_chans
+    unmixing_matrix$Component <- paste0("Comp", 1:pca) # names(data$signals)
   } else {
 
-    if (!requireNamespace("ica", quietly = TRUE)) {
-      stop("Package \"ica\" needed to use infomax or fastica. Please install it.",
-           call. = FALSE)
-    }
+    if (method == "sobi") {
+      if (!requireNamespace("JADE", quietly = TRUE)) {
+        stop("Package \"JADE\" needed to use SOBI. Please install it.",
+             call. = FALSE)
+      }
 
-    if (rank_check < ncol(data$signals)) {
-      warning(paste("Data is rank deficient. Detected rank", rank_check))
-    }
-
-    if (method == "fastica") {
-      ICA_out <- ica::icafast(data$signals,
-                              rank_check,
-                              maxit = maxit,
-                              tol = tol)
-    } else if (method == "infomax") {
-      ICA_out <- ica::icaimax(data$signals,
-                              rank_check,
-                              maxit = maxit,
-                              fun = "ext",
-                              tol = tol)
+      message("Running SOBI ICA.")
+      ICA_out <- sobi_ICA(data,
+                          maxiter = maxit,
+                          tol = tol,
+                          pca = pca)
     } else {
-      stop("Unknown method; available methods are sobi, fastica, and infomax.")
+
+      if (!requireNamespace("ica", quietly = TRUE)) {
+        stop("Package \"ica\" needed to use infomax or fastica. Please install it.",
+             call. = FALSE)
+      }
+
+      if (method == "fastica") {
+        message("Running fastica (ica).")
+        ICA_out <- ica::icafast(data$signals,
+                                nc = rank_check,
+                                maxit = maxit,
+                                tol = tol)
+      } else if (method == "infomax") {
+        message("Running extended-Infomax (ica).")
+        ICA_out <- ica::icaimax(data$signals,
+                                nc = rank_check,
+                                maxit = maxit,
+                                fun = "ext",
+                                tol = tol)
+      }
     }
 
-    ICA_out$S <- as.data.frame(ICA_out$S)
-    names(ICA_out$S) <- paste0("Comp", 1:ncol(ICA_out$S))
-    mixing_matrix <- as.data.frame(ICA_out$M)
-    names(mixing_matrix) <- paste0("Comp", 1:ncol(ICA_out$S))
-    mixing_matrix$electrode <- names(data$signals)
-    unmixing_matrix <- as.data.frame(t(ICA_out$W))
-    names(unmixing_matrix) <- paste0("Comp", 1:ncol(ICA_out$S))
-    unmixing_matrix$electrode <- names(data$signals)
+      ICA_out$S <- as.data.frame(ICA_out$S)
+      names(ICA_out$S) <- paste0("Comp", 1:ncol(ICA_out$S))
 
-    ica_obj <- eeg_ICA(mixing_matrix = mixing_matrix,
-                       unmixing_matrix = unmixing_matrix,
-                       signals = ICA_out$S,
-                       timings = data$timings,
-                       events = data$events,
-                       chan_info = data$chan_info,
-                       srate = data$srate,
-                       epochs = data$epochs,
-                       algorithm = method)
-    }
+      if (pca_flag) {
+        mixing_matrix <- as.data.frame(pca_decomp[, 1:pca] %*% ICA_out$M)
+        unmixing_matrix <- as.data.frame(MASS::ginv(as.matrix(mixing_matrix), tol = 0))
+        names(mixing_matrix) <- paste0("Comp", 1:pca)
+        names(unmixing_matrix) <- orig_chans
+        mixing_matrix$electrode <- orig_chans
+        unmixing_matrix$Component <- paste0("Comp", 1:pca)
+      } else {
+        mixing_matrix <- as.data.frame(ICA_out$M)
+        names(mixing_matrix) <- paste0("Comp", 1:ncol(ICA_out$M))
+        mixing_matrix$electrode <- names(data$signals)
+        unmixing_matrix <- as.data.frame(ICA_out$W)
+        names(unmixing_matrix) <- names(data$signals)
+        unmixing_matrix$Component <- paste0("Comp", 1:ncol(ICA_out$S))
+      }
+
+
+  }
+
+  ica_obj <- eeg_ICA(mixing_matrix = mixing_matrix,
+                     unmixing_matrix = unmixing_matrix,
+                     signals = ICA_out$S,
+                     timings = data$timings,
+                     events = data$events,
+                     chan_info = data$chan_info,
+                     srate = data$srate,
+                     epochs = data$epochs,
+                     algorithm = method)
   ica_obj
 }
 
@@ -96,12 +157,16 @@ run_ICA.eeg_epochs <- function(data,
 #'
 #' @param data Data to be ICAed.
 #' @param maxiter Maximum number of iterations of the joint diagonalization
+#' @param tol convergence tolerance.
+#' @param pca Number of PCA components.
 #' @author A. Belouchrani and A. Cichocki. Adapted to R by Matt Craddock
 #'   \email{matt@@mattcraddock.com}
 #' @keywords internal
 
 sobi_ICA <- function(data,
-                     maxiter) {
+                     maxiter,
+                     tol,
+                     pca) {
 
   n_epochs <- length(unique(data$timings$epoch))
   n_channels <- ncol(data$signals)
@@ -114,23 +179,16 @@ sobi_ICA <- function(data,
   ## Pre-whiten the data using the SVD. zero-mean columns and get SVD. NB:
   ## should probably edit this to zero mean *epochs*
   #do by epochs
-  amp_matrix <- split(data$signals, data$timings$epoch)
-  amp_matrix <- lapply(amp_matrix,
-                         function(x) sweep(x,
-                                           2,
-                                           Matrix::colMeans(x))
-                         )
-  # # amp_matrix <- lapply(amp_matrix,
-  # #                      function(x) scale(x, scale = FALSE))
-   amp_matrix <- as.matrix(do.call("rbind", amp_matrix))
-  #amp_matrix <- sweep(data$signals, 2, Matrix::colMeans(data$signals))
-  SVD_amp <- svd(amp_matrix)
+
+  # centre each epoch on zero
+  data <- rm_baseline(data, verbose = FALSE)
+
+  SVD_amp <- svd(data$signals)
 
   ## get the psuedo-inverse of the diagonal matrix, multiply by singular
   ## vectors
-   Q <- MASS::ginv(diag(SVD_amp$d), tol = 0) %*% t(SVD_amp$v) # whitening matrix
-   amp_matrix <- Q %*% t(amp_matrix)
-  #amp_matrix <- amp_matrix %*% SVD_amp$v
+  Q <- tcrossprod(MASS::ginv(diag(SVD_amp$d), tol = 0), SVD_amp$v) # whitening matrix
+  amp_matrix <- tcrossprod(Q, as.matrix(data$signals))
 
   ## reshape to reflect epoching structure
   dim(amp_matrix) <- c(n_channels,
@@ -162,8 +220,14 @@ sobi_ICA <- function(data,
   }
 
   epsil <- 1 / sqrt(N) / 100
+
+  if (epsil > tol) {
+    message("Setting tolerance to ", round(epsil, 4))
+    tol <- epsil
+  }
+
   V <- JADE::rjd(t(M),
-                 eps = epsil,
+                 eps = tol,
                  maxiter = maxiter)$V
 
   ## create mixing matrix for output
@@ -173,7 +237,10 @@ sobi_ICA <- function(data,
   # rescale vecs
   scaling <- sqrt(colMeans(mixing_matrix^2))
 
-  unmixing_matrix <- sweep(unmixing_matrix, MARGIN = 1, scaling, `*`) # scaled weights
+  unmixing_matrix <- sweep(unmixing_matrix,
+                           MARGIN = 1,
+                           scaling,
+                           `*`) # scaled weights
 
   mixing_matrix <- MASS::ginv(unmixing_matrix %*% diag(ncol(unmixing_matrix)),
                               tol = 0)
@@ -181,34 +248,38 @@ sobi_ICA <- function(data,
   dim(amp_matrix) <- c(n_channels,
                        n_times * n_epochs)
 
-  S <- unmixing_matrix %*% t(as.matrix(data$signals))
+  S <- tcrossprod(unmixing_matrix, as.matrix(data$signals))
   S <- as.data.frame(t(S))
   names(S) <- paste0("Comp", 1:ncol(S))
 
-  mixing_matrix <- as.data.frame(mixing_matrix)
-  names(mixing_matrix) <- paste0("Comp", 1:ncol(S))
-  mixing_matrix$electrode <- names(data$signals)
-  unmixing_matrix <- as.data.frame(unmixing_matrix)
-  names(unmixing_matrix) <- paste0("Comp", 1:ncol(S))
-  unmixing_matrix$electrode <- names(data$signals)
 
-  ica_obj <- eeg_ICA(mixing_matrix = mixing_matrix,
-                     unmixing_matrix = unmixing_matrix,
-                     signals = S,
-                     timings = data$timings,
-                     events = data$events,
-                     chan_info = data$chan_info,
-                     srate = data$srate,
-                     epochs = data$epochs,
-                     algorithm = "sobi")
-  ica_obj
+  list(M = mixing_matrix,
+       W = unmixing_matrix,
+       S = S)
 }
 
+
+#' Recreate channel timecourses from ICA decompositions.
+#'
+#' This function can be used to either recreate "mixed" (i.e. channel level)
+#' timecourses from an ICA decomposition, or to apply a set of ICA weights to a
+#' given dataset for the purpose of removing specific ICA components from that
+#' dataset.
+#'
+#' @param data An \code{eeg_ICA} or \code{eeg_epochs} object.
+#' @param ... Other parameters.
+#' @author Matt Craddock \email{matt@@mattcraddock.com}
+#' @export
 apply_ica <- function(data, ...) {
   UseMethod("apply_ica", data)
 }
 
-apply_ica.eeg_ICA <- function(data, comps, ...) {
+#' @param comps Components to remove.
+#' @describeIn apply_ica From given \code{eeg_ICA} object, recreate channel timecourses.
+#' @export
+apply_ica.eeg_ICA <- function(data,
+                              comps = NULL,
+                              ...) {
   ncomps <- ncol(data$mixing_matrix)
   new_mixmat <- data$mixing_matrix[1:(ncomps - 1)]
   new_mixmat[ , comps] <- 0
@@ -217,31 +288,46 @@ apply_ica.eeg_ICA <- function(data, comps, ...) {
   names(new_dat) <- data$chan_info$electrode
   out <-
     eeg_data(new_dat,
-             data$srate,
-             data$events,
-             data$chan_info,
-             data$timings,
-             data$continuous,
-             reference = NULL
-          )
+             srate = data$srate,
+             events = data$events,
+             chan_info = data$chan_info,
+             timings = data$timings,
+             reference = NULL,
+             epochs = NULL)
   class(out) <- c("eeg_epochs", "eeg_data")
   out
 }
 
-apply_ica.eeg_epochs <- function(data, decomp, comps, ...) {
+#' @param decomp An \code{eeg_ICA} object.
+#' @describeIn apply_ica Combine a specific set of ICA weights with any \code{eeg_epochs} object.
+#' @export
+apply_ica.eeg_epochs <- function(data,
+                                 decomp,
+                                 comps,
+                                 ...) {
   if (!is.eeg_ICA(decomp)) {
     stop("An eeg_ICA object must be supplied as decomp.")
-  } else {
-    ncomps <- ncol(decomp$mixing_matrix) - 1
-    new_mixmat <- decomp$mixing_matrix[1:(ncomps)]
-    new_mixmat[, comps] <- 0
-    source_sigs <- as.matrix(data$signals) %*%
-      t(as.matrix(decomp$unmixing_matrix[, 1:ncomps]))
-
-    new_dat <- as.matrix(new_mixmat) %*% t(source_sigs)
-    new_dat <- as.data.frame(t(new_dat))
-    names(new_dat) <- data$chan_info$electrode
-    data$signals <- new_dat
-    data
   }
+
+  ncomps <- ncol(decomp$mixing_matrix) - 1
+  nelecs <- ncol(data$signals)
+  sources <- unmix(data$signals,
+                   decomp$unmixing_matrix[, 1:nelecs])
+  keep_comps <- names(decomp$mixing_matrix[, 1:ncomps])[-comps]
+  new_sigs <- mix(sources[, -comps],
+                  decomp$mixing_matrix[, keep_comps])
+  data$signals <- tibble::as_tibble(new_sigs)
+  names(data$signals) <- data$chan_info$electrode
+  data
+}
+
+
+mix <- function(sources,
+                mixing_matrix) {
+  t(tcrossprod(as.matrix(mixing_matrix), as.matrix(sources)))
+}
+
+unmix <- function(data,
+                  unmixing_matrix) {
+  t(as.matrix(unmixing_matrix) %*% t(as.matrix(data)))
 }